+++ src/composables/index.ts
@@ -4,7 +4,6 @@
 
 /** Barrel re-exporting every Vue 3 composable. */
 export * from "./useAppLockState";
-export * from "./useClipboardNotify";
 export * from "./useCommitSignature";
 export * from "./useDivergence";
 export * from "./useLockActivity";